updated Order, orderItem in 21.05.2022

+++ src/vn/mvpthinh/services/IOrderItemService.java
@@ -1,16 +1,16 @@
 package vn.mvpthinh.services;
 
-import vn.mvpthinh.model.Order;
 import vn.mvpthinh.model.OrderItem;
 
 import java.util.List;
 
 public interface IOrderItemService {
+
     List<OrderItem> findAll();
 
     void add(OrderItem newOrderItem);
 
     void update(OrderItem newOrderItem);
 
-    OrderItem getOrderItemById(int id);
+    OrderItem getOrderItemById(Long id);
 }
